\name{achievscore}
\alias{achievscore}


\non_function{}
\title{Achievement in training}
\usage{data(achievscore)}
\description{
  Experiment on the effect of three training methods
\cr
Analysis of Covariance
}
\format{
   Data frame containing
\describe{
    \item{method:}{Indicator for training method: A, B or C}
    \item{y:}{achievement score}
    \item{x:}{aptitude score, measured at the start of the experiment}
  }
}
\details{
Data are orginally from Winer,  B.J. (1971), Statistical Principles in
Experimental Design,
Second edition, McGraw Hill, New York
}
\source{
Dobson, A. (2002) \emph{An introduction to Generalized Linear Models},
London: Chapman \& Hall, p. 106, table 6.12
        }
\references{
}
\examples{
data(achievscore)
attach(achievscore)


}





\keyword{datasets}
