\name{activityCowLogger}
\alias{activityCowLogger}
\docType{data}
\title{Daily activity of cows.}
\description{
The cows' id for the logger used for daily activity measurments of the cows.
A subset of the data \code{\link{activityCowsLong}}.
}
\usage{data(activityCowLogger)}
\format{
A dataframe with  2 variables.
\describe{
\item{cow:}{id of cow}
\item{logger:}{id of logger}
}}

\source{
}

\seealso{\code{\link{activityCowsLong}}}

\references{

}




\examples{
data(activityCowLogger)  
}
\keyword{datasets}	




