\name{activityCowsLong}
\alias{activityCowsLong}
\docType{data}
\title{Daily activity of cows.}
\description{
Twenty loggers placed on twenty cows transmit automatically daily activity measures to a PC. 
The data contain additional information about the farms on which the measurements were taken.
}
\usage{data(activityCowsLong)}
\format{
A dataframe with  7 variables.
\describe{
\item{logger:}{id of logger}
\item{cow:}{id of cow}
\item{farm:}{id of farm}
\item{owner:}{Owner of the farm}
\item{type:}{Production type of the farm}
\item{day:}{Day of measurement}
\item{activity:}{Actitvity measurement}
}}

\source{
}

\seealso{\code{\link{activityFarmCow}}}

\references{

}




\examples{
data(activityCowsLong)  
}
\keyword{datasets}	




