\name{activityFarmCow}
\alias{activityFarmCow}
\docType{data}
\title{Daily activity of cows.}
\description{
Information on farms on cows for which daily activity measurements were taken.
A subset of the data \code{\link{activityCowsLong}}.
}
\usage{data(activityFarmCow)}
\format{
A dataframe with  2 variables.
\describe{
\item{farm:}{id of farm}
\item{cow:}{id of cow}
}}

\source{
}

\seealso{\code{\link{activityCowsLong}}}

\references{

}




\examples{
data(activityFarmCow)  
}
\keyword{datasets}	




