\name{activityFarmInfo}
\alias{activityFarmInfo}
\docType{data}
\title{Daily activity of cows.}
\description{
Information on farms  on which  daily activity measurements on cows were taken.
A subset of the data \code{\link{activityCowsLong}}.
}
\usage{data(activityFarmInfo)}
\format{
A dataframe with  2 variables.
\describe{
\item{farm:}{id of farm}
\item{owner:}{owner of farm}
\item{type:}{production type of farm}
}}

\source{
}

\seealso{\code{\link{activityCowsLong}}}

\references{

}




\examples{
data(activityFarmInfo)  
}
\keyword{datasets}	




