\name{activityLogger}
\alias{activityLogger}
\docType{data}
\title{Daily activity of cows.}
\description{
Twenty loggers where placed on twenty cows and on three days the cows'  daily activity was measured.
A subset of the data \code{\link{activityCowsLong}}.
}
\usage{data(activityLogger)}
\format{
A dataframe with  4 variables.
\describe{
\item{logger:}{id of logger}
\item{day1:}{activity measurement for day 1}
\item{day2:}{activity measurement for day 2}
\item{day3:}{activity measurement for day 3}
}}

\source{
}

\seealso{\code{\link{activityCowsLong}}}

\references{

}




\examples{
data(activityLogger)  
}
\keyword{datasets}	




