\name{allergy}
\alias{allergy}
\docType{data}
\non_function{}
\title{Allergic reaction of patients}
\usage{data(allergy)}
\description{
357 pateinets with a diagnosis of hay fever were tested for slleroc reaktion for
extracts from pollen, animals and mites}
\format{
   Data frame containing
  \describe{
    \item{pollen:}{0: no reaction; 1: reaction}
	\item{animal:}{0: no reaction; 1: reaction}
	\item{mites:}{0: no reaction; 1: reaction}
	\item{count:}{count of patients}
}}
\details{
}
\source{ B. Madsen (1982)}
\references{
Anderssen, Søren (1990) Analyse af tælledata, Institut for Matematik og Fysik, KVL,
Thorvaldsensvej 40, 1871 Frederiksberg C, 2. oplag, side 36
København
}


\examples{
data(allergy)
}





\keyword{datasets}
