\name{anther}
\alias{anther}
\docType{data}
\title{Embryogenic anthers.}
\description{
Anthers of the plant species Datura innoxia were prepared under different conditions and the number of embryogenic anthers counted.
}
\usage{data(anther)}
\format{
A dataframe with  4 variables.
\describe{
\item{y:}{number of embryogentic anthers}
\item{n:}{number of anthers prepared}
\item{storage:}{storage conditions}
\item{force:}{Centrifuging force}
}
}

\source{
}

\references{
Dobson, Annette, (2002), An Introduction to Generalized Linear Models, Chapma and Hall
}




\examples{
data(anther)  
}
\keyword{datasets}	




