\name{bachelor_avl}
\alias{bachelor_avl}


\non_function{}
\title{Breeding value of bulls}
\usage{data(bachelor_avl)}
\description{
  The data gives the breeding values for 20 bulls from 2 different
   breeders.
  A bull with breeding value above above 100 is usually
  considerd to be better than the average bull.
    }
\format{
   Data frame containing
\describe{
    \item{avler:}{Number 1 and 2 to indicate the breeder}
    \item{value:}{the breeding value of a bull}
  }
}
\source{
  Ficticious data
     }
\examples{
data(bachelor_avl)
}

\keyword{datasets}
