\name{bactsucrose}
\alias{bactsucrose}
\docType{data}
\non_function{}
\title{Influence of nutritients on growth of bacteria}
\usage{data(bactsucrose)}

\description{
In an experiment on nutrients effects on bacterial growth,
bacteria were grown on agar plates in solutions with different concentrations of the nutrients sucrose (4 levels) and 
leucine (3 levels).
The experiment was replicated on 4 days and on each day all 12 treatment combinations were applied.
The numbers of bacteria in the solution were determined.
}
\format{
   Data frame containing
  \describe{day sucrose leucine density
	    \item{day:}{day of experiment}
    \item{sucrose:}{level of scurose nutrient}
    \item{leucine:}{level of leucine nutrient}
    \item{density:}{number fo bacteria}
}
}
\details{
}
\source{
Grafen, A. and Hails, R. (2002), \emph{Modern Statistics for the Life Sciences}, Oxford University Press, London, chapter 9.4, p. 169
}

\references{
}


\examples{
data(bactsucrose)
}





\keyword{datasets}
