\name{BarleyYield}
\alias{BarleyYield}


\non_function{}
\title{BarleyYield}
\usage{data(BarleyYield)}
\description{
  Experiment on the effect of three training methods
\cr
Analysis of Covariance
}
\format{
   Data frame containing
\describe{
    \item{Nitrogen:}{}
    \item{Water:}{}
    \item{Block:}{}
    \item{Yield:}{}
  }
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(BarleyYield)
}





\keyword{datasets}
