\name{BeefScores}
\alias{BeefScores}


\non_function{}
\title{BeefScores}
\usage{data(BeefScores)}
\description{
  Experiment on the effect of three training methods
\cr
Analysis of Covariance
}
\format{
   Data frame containing
\describe{
    \item{rep:}{}
    \item{block:}{}
    \item{treat:}{}
    \item{score:}{}
  }
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(BeefScores)
}





\keyword{datasets}
