\name{bolger}
\alias{bolger}
\docType{data}
\non_function{}

\title{Rodents in habitat fragments}
\usage{data(bolger)}
\description{
Prence/absence of any specy of native rodent (except Microtus califcornicus) dependent on
distance to the nearest source canyon, age since the fragment was isolated by urbanization and the percentage of fragment 
area covered in shrubs (bushes).
}
\format{
  This data frame contains:
  \describe{
       \item{pershrub:}{percentage of fragment area covered in shrubs}
       \item{distx:}{distance [m] to the nearest source canyon}
       \item{age:}{age [years] since fragment was isolated by urbanization}
       \item{rodentspa:}{presence/absence of native rodent species [1=present, 0=absent]}
  }}
\details{
}
\source{
        }
\references{
Quinn, Gerry. P. and Keough, Michael J. (2002) \emph{Experimental Design and Data Analysis for Biologists}, University Press, Cambridge
Box 13-2, Fig 13-2
}

\examples{
data(bolger)
}
\keyword{datasets}
