\name{broadbalk}
\alias{broadbalk}
\docType{data}
\non_function{}

\title{broadbalk}
\usage{data(broadbalk)}
\description{
The yields of dressed grain were obtained from two plots on Broadwalk
wheat field
between the years 1855 and 1884. The only difference in manurial
treatment was one plot (A) received nitrate of soda
the other plot (B) an equivalent quantity of nitrogen as sulphate of
ammonia.
In the course of the experiment the A plot appears to gain in yield
on plot B.
}
\format{
  This data frame contains:
  \describe{
       \item{year:}{years 1855-1884}
       \item{A:}{yield [bushels per acre], plot treated with
	 sodium nitrate (NaNO3)}
        \item{B:}{yield [bushels per acre], plot treated with 
	ammonium sulfate ((NH4)2SO4) }
  }}
\details{linear regression, curvilinear regression}
\source{
}

\references{
  Fisher, R. A. (1936) Statistical Methods for Research Workers
  Oliver and Boyd, 6th ed., Table 29 of Example 22, pp. 142--146.
}

\examples{
}
\keyword{datasets}
