\name{butterfat}
\alias{butterfat}
\docType{data}
\non_function{}

\title{butterfat}
\usage{data(butterfat)}
\description{
Average  butterfat percentage for random samples of 
mature cows (older than 5 years)
and 2 year old cows from five different breeds.
}
\format{
  This data frame contains:
  \describe{
       \item{fat:}{Butter fat percentage}
       \item{age:}{Age level: mature (> 5 years, 2-years: 2 years old)}
       \item{breed:}{Five breeds:
Ayrshire, Canadian, Guernsey, Holstein-Fresian,  Jersey }
}}
\details{ANOVA, 2-way}
\source{
Sokal, R. and Rohlf, F.J. (1981) \emph{Biometry}, 2nd edition,
p. 368. 
}

\references{
}

\examples{	
data(butterfat)
}
\keyword{datasets}
