\name{BygYield}
\alias{BygYield}


\non_function{}
\title{BygYield}
\usage{data(BygYield)}
\description{
}
\format{
   Data frame containing
\describe{
    \item{Nitrogen:}{}
    \item{Water:}{}
    \item{Block:}{}
    \item{yield:}{}
  }
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(BygYield)
}





\keyword{datasets}
