\name{carrot}
\alias{carrot}
\docType{data}
\title{Measurements of carrots}
\description{
The weight of 144 carrots 
  \cr skewed distribution
}
\usage{data(carrot)}
\format{
A dataframe with 1 variable. 
\describe{
\item{weight:}{weight of carrot in g}
}
}

\source{
Mead, R, Curnow, R.N. and Hasted, A.M. (1993) Statistical methods in Agriculture and 
Experimental Biology Chapman and Hall,  Table 2.3
}



\examples{
data(carrot)
}

\keyword{datasets}




