\name{caturin}
\alias{caturin}
\docType{data}
\non_function{}
\title{Disease of the urinary tract of cats}
\usage{data(caturin)}
\description{
In a an 
}
\format{
   Data frame containing
  \describe{
    \item{count:}{number of persons asked; one cat per person}
    \item{infect:}{yes; no}
    \item{motion:}{0: little motion, 1: a good deal}
    \item{dryfeed:}{0: a bit dryfeed; 1:mainly dryfeed}
}}
\details{
}
\source{
}
\references{
Anderssen, Søren (1990) Analyse af tælledata, Institut for Matematik og Fysik, KVL,
Thorvaldsensvej 40, 1871 Frederiksberg C, 2. oplag, side 44
København
}


\examples{
data(caturin)
}





\keyword{datasets}
