\name{cheeseTaste}
\alias{cheeseTaste}
\docType{data}
\non_function{}
\title{taste of four different cheese}
\description{
Persons were randomly assigned to four different types of cheeses and asked to
characterize the taste on a scale from 1 to 9
}
\usage{data(cheeseTaste)}
\format{
A dataframe with 3 variables. 
\describe{
\item{taste}{response: evaluation of taste:  from 1: strong dislike to 9: excellent taste}
\item{cheese:}{ type of cheese tasted}
\item{count:}{ number of answers}
}
}


\source{
McCullagh, P.  and  Nelder, J. A. (1989) \emph{Generalized Linear Models}, London, U.K.: Chapman & Hall
        }



\examples{}


\keyword{datasets}




