\name{cherrytree}
\alias{cherrytree}
\non_function{}
\title{volum, height and diameter of cherry trees}
\usage{data(cherrytree)}
\description{
measurements of the girth(=diam), height and
     volume of timber in 31 felled black cherry trees.  Note that girth
     is the diameter of the tree (in inches) measured at 4 ft 6 in
     above the ground.
}
\format{
  This data frame contains the following columns:
\describe{
    \item{diam:}{diameter of tree in inch [2.54 cm] }
    \item{height:}{height of tree in feet [30.48 cm]}
    \item{vol:}{volume of tree in cubic feet}
  } 
}
\details{
}

\source{
Atkinson, A. C. (1982) Regression diagnostics, transformations and constructed variables (with discussion). Journal of the Royal Statistical Society, Series B, 44, 1-36.
}

\references{
Ryan, T. A. Jr., Joiner, B. L. and Ryan, B. F. (1985) The Minitab Student Handbook, Boston: Duxbury Press, 328-329. 
}

\examples{
data(cherrytree)
#
}
\keyword{datasets}
