\name{cowZgeno}
\alias{cowZgeno}
\docType{data}
\non_function{}
\title{Allergic reaction of patients}
\usage{data(cowZgeno)}
\description{
357 pateinets with a diagnosis of hay fever were tested for slleroc reaktion for
extracts from pollen, animals and mites}
\format{
   Data frame containing
  \describe{
    \item{race:}{breed of cows: RDM, SDM and jersey}
	\item{type:}{gwenotype: ZZ, Zz or zz}
	\item{count:}{count of cows}
}}
\details{
}
\source{ Neimann-Sørensen (1958)}
\references{
Anderssen, Søren (1990) Analyse af tælledata, Institut for Matematik og Fysik, KVL,
Thorvaldsensvej 40, 1871 Frederiksberg C, 2. oplag, side 2
København
}

\examples{
data(cowZgeno)
}





\keyword{datasets}
