\name{CrimeRate}
\alias{CrimeRate}
\docType{data}
\non_function{}

\title{CrimeRate}
\usage{data(CrimeRate)}
\description{
Crime rates per 100,000 inhabitants in states of the USA for different
crime types.
}
\format{
  This data frame contains:
  \describe{
       \item{State:}{State of the USA}
       \item{Murder:}{crime of murder}
       \item{Rape:}{}
       \item{Robbery:}{}
       \item{Assault:}{}
       \item{Burglary:}{residential theft}
       \item{Larceny:}{unlawful taking of personal property (pocket picking)}
       \item{AutoTheft:}{}
}}
\details{
PCA
}
\source{
SHD
}

\references{
}

\examples{	
data(CrimeRate)
}
\keyword{datasets}
