\name{degroten}
\alias{degroten}
\docType{data}
\non_function{}
\title{Two insecticides probed on insects}
\usage{data(degroten)}
\description{
Insects are exposed to different doses of two  insectiocided rotenone and deguelin}
\format{
   Data frame containing
  \describe{
	\item{insect:}{insecticide: rotenone or deguelin}
    \item{dose:}{dose of insecticide}
	\item{deaths:}{Number of insects killed}
	\item{total:}{Number of insects exposed}
}}
\details{Logistic regression}
\source{Exercise 3.6 (6) Lindsey, J. K. (1995) Modelling Frequency and count data, Oxford Science publications}
\references{
}

\examples{
data(degroten)
}





\keyword{datasets}
