\name{fertilhypo}
\alias{fertilhypo}
\docType{data}
\non_function{}

\title{Fertilhypo}
\usage{data(fertilhypo)}
\description{
Constructed dataset about the yield of three fertilisers.
There are three experiments. In the first the mean of the three
fertilisers is 2, 5 and 9 and the residual variane is 0.36.
 In the iother tweo experiment the fertilisers have all mean 5, but 
the error variance is 5.29 and 0.36.
}
\format{
  This data frame contains:
  \describe{
       \item{fertil:}{fertilsers 1, 2 and 3}
     \item{yA:}{yield,  
          mean for all fertilisers 5, population error variance 5.29}
     \item{yB:}{yield,  
          mean for all fertilisers 5, population error varianece 0.36}
	 \item{yC:}{yield,  
          mean per fertiliser is 2, 5 and 9, population error variance 0.36}
       
}}
\details{
One way Anova, constructed data to show effect of different residual variances}
\source{
}

\references{
}

\examples{	
data(fertilhypo)
#data generated by
set.seed(8)
treat  <-  rep(c(1,  2,  3),  each  = 10)
##treat  <-  rep(c(1,  2,  3),  each  =  c(10,  11,  12))

x  <-  1:length(treat)
treatf  <-  factor(treat)
yA  <-  5  +  rnorm(x,  sd  =  2.3)
yA  <-  ifelse(yA  <  0,  0.55,  yA)
yB  <-  5  +  rnorm(x,  sd  =  0.6)
yC  <-  c(2,  5,  9)[treatf]  +  rnorm(x,  sd  =  0.6)
}
\keyword{datasets}
