\name{flagstone}
\alias{flagstone}
\docType{data}
\non_function{}

\title{flagstone}
\usage{data(flagstone)}
\description{
Flagstones a priced fromconcrete.
A manufacturer was interetses, wthther additives 'a', 'b' or a mixture of them, could imporve the stability of the flaghstone.

From a ppduciton of flagstones with the fo

Experiment to analyse the influence of meachanical stress on the growth of 
soya bean plants.
52 seedlings were potted, one seedling per plot.
The seedlings were randomly splitted up into four groups of
13 plants.
In two groups the plants were shaken for 20 minutes daily.
The groups that received the same amount of stress were grown under two different light conditions.
After 16 days the plants were harvested and the total area of leaves
for each plant was measured.
}
\format{
  This data frame contains:
  \describe{
       \item{a:}{Additive a: -1 not added, 1 : 10 kg of additive a added}
       \item{b:}{Additive b: -1 not added, 1 : 10 kg of additive a added}
       \item{strength:}{Compression strength of stone on maga pascal}
}}
\details{
}
\source{
Blæsild, Preben and Grandfeldt, Jørgen (2003) 
\emph{Statistics with applications in
Biology and Geology}, Chapman and Hall,
Exercise 4.8, pp. 239. 
}

\references{
}

\examples{	
data(flagstone)
}
\keyword{datasets}
