\name{FungiToVar}
\alias{FungiToVar}
\docType{data}
\non_function{}
\title{Fungicides}
\usage{data(FungiToVar)}
\description{Fungicides}
\format{
   Data frame containing
  \describe{
	\item{Rep:}{replication of block}
    \item{Variety:}{varieties; 1,2, 3 or 4}
	\item{Fungicide:}{fingicide; a,b,c,d,e or f}
	\item{Yield:}{yield}
	}}
\details{Split plot design}
\source{}
\references{Kristian Kristensen}
\examples{
data(FungiToVar)
}





\keyword{datasets}
