\name{hubble}
\alias{hubble}
\docType{data}
\non_function{}

\title{Hubble}
\usage{data(hubble)}
\description{
Hubbles data published in 1929 on the distance of a galaxy from ours and 
its velocity with which it moves away from the milkyway.
}
\format{
  This data frame contains:
  \describe{
       \item{distance:}{distance in 3.26 *1000000 light years}
       \item{velocity:}{km/second}
}}
\details{
}
\source{
}

\references{
Jens Ledet Jensen (2009) Et Nanokurs i Statistik, chapter 9, p. 96.
}

\examples{	
data(hubble)
}
\keyword{datasets}
