\name{mealy}
\alias{mealy}


\non_function{}
\title{mealy}
\usage{data(mealy)}
\description{
}
\format{
   Data frame containing
\describe{
    \item{treat:}{}
    \item{plant:}{}
    \item{change:}{}
}
}
	\details{
}
\source{
        }
\references{Example 13.1 from Oehlert, page 316
Oehlert 
}
\examples{
data(mealy)
}





\keyword{datasets}
