\name{nematodes}
\alias{nematodes}


\non_function{}
\title{Number of nematodes}
\usage{data(nematodes)}
\description{
Number of nematodes in 15 samples of solutions of 20 \eqn{\mu}l and 40 
\eqn{\mu}l.
}
\format{
  This data frame contains 
  \describe{
    \item{volume:}{volume of solution \eqn{\mu}l}
    \item{count:}{number of nematodes counted in a sample solution}
    \item{freq:}{frequency of observations of the specific number of nematodes}
  }
}
\details{
}
\source{
 Lars Stubsgaard, Institute of Biological Sciences, University of Arhus
        }
\references{
Blæsild, P. and Granfeldt, J. (2003) \emph{Statistics with applications in Biplogy and Geography}, Chapman and Hall-CRC: London, p. 358
}
\examples{
data(nematodes)

}
\keyword{datasets}
