\name{NtilHavre}
\alias{NtilHavre}


\non_function{}
\title{NtilHavre}
\usage{data(NtilHavre)}
\description{
}
\format{
   Data frame containing
\describe{
    \item{Variety:}{}
    \item{Nitrogen:}{}
    \item{Yield:}{}
    \item{Block:}{}
  }
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(NtilHavre)
}





\keyword{datasets}
