\name{ovulation}
\alias{ovulation}
\docType{data}
\title{Ovulation rate for sows}
\description{
An experiment about the influcence 
of four feed suplements  PG600, PMSG, FSH or saline
on the ovulation rate of sows.

20 sows were randomly selected from a farm and assigned to five pens.
Treatments were randomly assigned to the four sows per pen.
  \cr
 RCBD
\cr
}
\usage{data(ovolution)}
\format{
A dataframe with 3 variables. 
\describe{
\item{pen:}{pen number}
\item{treat:}{4 treatments FSH, PG600, PMSG and saline}
\item{rate:}{ovulation rate}
}
}

\source{
Miroslav Kaps and  William Lamberson (2009) biostatistics for animal science
CABI, p. 335.
}



\examples{
data(ovulation)
}

\keyword{datasets}




