\name{photosyn}
\alias{photosyn}
\docType{data}
\non_function{}
\title{Radiation and photosynthesis}
\usage{data(photosyn)}
\description{
In a study the dependence of photosynthesis rate on radiation and water availability was examnined.
}
\format{
   Data frame containing
  \describe{
    \item{water:}{level of water availability}
    \item{radiation:}{amount of radiation }
    \item{rate:}{rate of photosynthesis}
  }
}
\details{
}
\source{
Krzanowski, W (1998) \emph{An Introduction to Statistical Modelling}. Arnold, London,
 p.107       }
\references{
}
\examples{
data(photosyn)
}





\keyword{datasets}
