\name{pigDat}
\alias{pigDat}


\non_function{}
\title{pigDat}
\usage{data(pigDat)}
\description{
  Experiment on the effect of three training methods
\cr
Analysis of Covariance
}
\format{
   Data frame containing
\describe{
    \item{Weight:}{Indicator for training method: A, B or C}
    \item{Treatment:}{achievement score}
    \item{B12:}{aptitude score, measured at the start of the experiment}
    \item{Antibiotic:}{aptitude score, measured at the start of the experiment}
  }
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(pigDat)
}





\keyword{datasets}
