\name{pigweight}
\alias{pigweight}
\docType{data}
\title{Growth of pigs.}
\description{
Growth curves of 48 pigs.
}
\usage{data(pigweight)}
\format{
A dataframe with  9 variables.
\describe{
\item{weeki:}{weight at week i}
}
}

\source{
}

\references{
Diggle, PJP, Heagerty KY, Liang, KY and Zeger, L. (2002) Analysis of Longitudinal Data, Oxford
}




\examples{
data(pigweight)  
}
\keyword{datasets}	




