\name{pistons}
\alias{pistons}


\non_function{}
\title{Failure of piston rings}
\usage{data(pistons)}
\description{
  These data concern the number of failures of piston rings for each of three legs of four steam-driven compressors. 
\cr
ANOVA, Randomised complete block design,  normal or poisson
}
\format{
   Data frame containing
  \describe{
    \item{pistons:}{Number of failures of piston rings}
    \item{leg:}{Position of leg (North=1, Centre=2, South=3)}
    \item{compressor:}{Compressor identification number}
  }
}
\source{
  Davies, O.L. and Goldsmith, P.L. (1972) Statistical Methods in Research and Production, 4th edn. Oliver and Boyd, Edinburgh, p. 324. }
\references{
}
\examples{
data(pistons)
}

\keyword{datasets}
