\name{poisoned}
\alias{poisoned}
\alias{poisonedWide}
\docType{data}
\title{Survial of poisoned animals}
\description{
  The survival of animals exposed to threee different poisons and
  four treatments is analysed.
    }
\usage{data(poisoned)}
\format{
  A data frame with 48 observations on 3  variables.
  \describe{
    \item{time}{survival time [10hr]}
    \item{treat}{type of treatment: A, B, C or D}
    \item{poison}{type of poison: 1,2 or 3}
    }
}
% \details{

% }
\source{
  Box, G. E. P. and Cox, D. R. (1964), An Analysis of Transformations,
  Journal of the Royal Statistical Society, 211-243, discussion 244-252.
   }
% \references{
%   ~~ possibly secondary sources and usages ~~
% }
\examples{
data(poisoned)
}
\keyword{datasets}
