\name{polis}
\alias{polis}
\docType{data}
\non_function{}

\title{Presence of Lizards on Island}
\usage{data(polis)}
\description{
Subaset of  data from a study on factors that influence the spider population on islands  in the Gulf of Califonia.
The present data give the presence/absence of the lizard (firben) uta on 19 islands in the Golf of California
and the ratio of the perimeter to the area of the islands.
}
\format{
  This data frame contains:
  \describe{
       \item{island:}{name of the island}
       \item{paratio:}{perimeter to area ratio of the islands}
       \item{uta:}{presence or absence of lizards (Uta) on the islands (P=present, A=absence)}
  }}
\details{
}
\source{
Polis, G.A., Hurd, S.D., Jackson, C.D. and Sanchez-Pinero, F. (1998) Multifactor
population limitation:
variable spatial and temporal control of spiders on Gulf of California islands,
Ecology, 79, pp. 490--502}

\references{
Quinn, Gerry. P. and Keough, Michael J. (2002) \emph{Experimental Design and Data Analysis for Biologists}, University Press, Cambridge
Box 13-1
}

\examples{
data(polis)
}
\keyword{datasets}
