\name{PotPlants}
\alias{PotPlants}


\non_function{}
\title{PotPlants}
\usage{data(PotPlants)}
\description{
}
\format{
   Data frame containing
\describe{
    \item{ID:}{}
    \item{Rep:}{}
    \item{Drain:}{}
    \item{SphType:}{}
	\item{PotType:}{}
	\item{DifWater:}{}
  }
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(PotPlants)
}





\keyword{datasets}
