\name{protein}
\alias{protein}


\non_function{}
\title{Diets with different contents of protein}
\usage{data(protein)}
\description{
   Six groups, each of ten rats, were fed on diets which differed according to source of protein and amount of protein 
in diet. The weight gain for each rat was recorded. 
\cr
ANOVA, 2-way,   
}
\format{
   Data frame containing
  \describe{
    \item{weight:}{Weight gain [g]}
    \item{protein:}{Protein source: beef, cereal, pork}
    \item{amount:}{Amount of protein: low, high}
  }
}
\source{
  Snedecor, G.W. and Cochran, G.C. (1967) Statistical Methods, 6th edn. Iowa State University Press, p. 347. 
}
\examples{
data(protein)
}

\keyword{datasets}
