\name{raindat}
\alias{raindat}
\docType{data}
\title{Rainfall in 2009}
\description{
  Rainfall  on irregular measured time interval in 2009.
}
\usage{data(raindat)}
\format{
A data frame 
\describe{
    \item{day}{day of measurement: character in the format 29-05-2009 (\code{\%d-\%m-\%Y})}
    \item{time}{time of measuremtn character in the cormat 24:17:01
      (\code{\%H:\%M:\%S)}}
    \item{rainfall}{precipitation in mm}
  }
}
\details{
  Exercise on date- time objects
}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  ~~ possibly secondary sources and usages ~~
}

\examples{
data(raindat)
}
\keyword{datasets}
