\name{Roses}
\alias{Roses}


\non_function{}
\title{Roses}
\usage{data(Roses)}
\description{
  Experiment on the effect of three training methods
\cr
Analysis of Covariance
}
\format{
   Data frame containing
\describe{
    \item{Grower:}{Indicator for training method: A, B or C}
    \item{Dealer:}{achievement score}
    \item{Customer:}{aptitude score, measured at the start of the experiment}
    \item{Block:}{aptitude score, measured at the start of the experiment}
    \item{DaysAntibiotic:}{aptitude score, measured at the start of the experiment}
  }
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(Roses)
}





\keyword{datasets}
