\name{Seedling}
\alias{Seedling}


\non_function{}
\title{Seedling}
\usage{data(Seedling)}
\description{
}
\format{
   Data frame containing
\describe{
    \item{Block:}{}
    \item{Iblock:}{}
    \item{N:}{}
    \item{P:}{}
	\item{K:}{}
	\item{Germin:}{}
}
}
	\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(Seedling)
}





\keyword{datasets}
