\name{soreThroat}
\alias{soreThroat}
\docType{data}
\title{Post-operative sore thoat.}
\description{
Influcence of method of anaesthetic gas application.
In a study the effect of  two different airway-masks on the post-operative
experience of sore throat was analysed. 
Data from surgeries with 10 consultants (anathesists) were collected 
on different patients.  
}
\usage{data(soreThroat)}
\format{
A dataframe with 4 variables. 
\describe{
\item{consultant}{anaethesist}
\item{airway}{type of aiway mas: FM: traditional face mask, LMA: laryngeal mask}
\item{y}{patients with sore throat whithin 24 hours after operation}
\item{n}{number of patients treated}
}
}

\source{
}

\references{
Collett, David (2003) Modelling Binary Data, Chapman and Hall, example 8.1,
p. 278}




\examples{
data(soreThroat)  
}
\keyword{datasets}




