\name{stork}
\alias{stork}
\docType{data}
\non_function{}
\title{stork}
\usage{data(stork)}
\description{
Number of hatching pairs of storks collected 1980-1990 for 17 European 
countries,
geographical information and the  population size and birth rate for humans.
}
\format{
  This data frame contains:
  \describe{
       \item{country:}{European country.}
       \item{area:}{area [\eqn{km^2}{km2}].}
       \item{nstork:}{Number of hatching storks pairs.}
       \item{pop:}{Human population [\eqn{10^6}].}
       \item{rate:}{Birth rate of human population (1000/year).}
}}
\details{Nonsense correlation}
\source{
Matthews, R. (2000) Storks deliver babies, (rho=0.008).
Teaching Statistics, 22(2), 36-38}
\references{
}

\examples{	
data(stork)
}
\keyword{datasets}
