\name{SugarBeets}
\alias{SugarBeets}


\non_function{}
\title{SugarBeets}
\usage{data(SugarBeets)}
\description{
}
\format{
   Data frame containing
\describe{
    \item{Rep:}{}
    \item{Block:}{}
    \item{Sucre:}{}
    \item{Variety:}{}
}	
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(SugarBeets)
}





\keyword{datasets}
