\name{teatlesion}
\alias{teatlesion}
\docType{data}
\non_function{}
\title{Teat lesions of  cows}
\usage{data(teatlesion)}
\description{
357 pateinets with a diagnosis of hay fever were tested for slleroc reaktion for
extracts from pollen, animals and mites}
\format{
   Data frame containing
  \describe{
    \item{level:}{upper or lower part of the teat}
	\item{pos1:}{right or left}
	\item{pos2:}{front or back}
	\item{count:}{count of glands}
}}
\details{
}
\source{ Agger and Hessselholt (1986)}
\references{
Anderssen, Søren (1990) Analyse af tælledata, Institut for Matematik og Fysik, KVL,
Thorvaldsensvej 40, 1871 Frederiksberg C, 2. oplag, page 2
København
}

\examples{
data(teatlesion)
}





\keyword{datasets}
