\name{titanium}
\alias{titanium}
\docType{data}
\title{extinction coefficient and  titanium content of cement}
\description{
Experiment on the relation between the titanium content of cement 
and the extinction coefficient of solutions of the cement
}
\usage{data(titanium)}
\format{
A dataframe with 3 variables. 
\describe{
\item{contents:}{per cent titanium contents of the cement}
\item{length:}{length of cuvette[cm]}
\item{extinction:}{extinction coefficient}
}
}

\source{
Hald, A. (1952) Statistical Theory with Engineering Applications,
New York: Wiley
}

\references{
}



\examples{
data(titanium)
attach(titanium)

}
\keyword{datasets}




