\name{turtle}
\alias{turtle}
\docType{data}
\non_function{}
\title{Hatching of turtle eggs}
\usage{data(turtle)}
\description{
The \code{turtle} data frame gives the number of male and female hatched eggs
of turtles under different temperatures.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{batch:}{number of batch}
    \item{temp:}{temperature of the batches under hatching}
    \item{male:}{number of eggs with male turtles}
    \item{female:}{number of eggs with female turtles}
    \item{state:}{location whre the eggs were collected}
  } 
}
\details{
}

\source{
K. Koehler, Iowa State Univeristy, see:\cr
\url{http://courses.ncssm.edu/math/Stat_Inst/PDFS/REG3_LOG.pdf}
}

\references{
}

\examples{
#data(turtle)
#
} 

\keyword{datasets}
