\name{watercontent}
\alias{watercontent}
\docType{data}
\title{Water content change under different storing conditions}
\description{
14 samples of a producte were randomly divided into 5 groups 
and each group was stored under different conditions.
After storage the water content was measured.
}
\usage{data(watercontent)}
\format{
A dataframe with 3 variables. 
\describe{
\item{group}{number of group}
\item{water}{water content in per cent}
}
}

\source{
Hald, A. (1952) Statistical Theory with Engineering Applications,
New York: Wiley, p. 432
}

\references{
}



\examples{
data(watercontent)
attach(watercontent)

}
\keyword{datasets}




