\name{WearTesting}
\alias{WearTesting}


\non_function{}
\title{WearTesting}
\usage{data(WearTesting)}
\description{
  Experiment on the effect of three training methods
\cr
Analysis of Covariance
}
\format{
   Data frame containing
\describe{
lock Treat Pos Loss
    \item{Block:}{Indicator for training method: A, B or C}
    \item{Treat:}{achievement score}
    \item{Pos:}{aptitude score, measured at the start of the experiment}
    \item{Loss:}{aptitude score, measured at the start of the experiment}
  }
}
\details{
}
\source{
        }
\references{ Kristian Kristensen
}
\examples{
data(WearTesting)

}





\keyword{datasets}
