\name{whitebloodcell}
\alias{whitebloodcell}
\docType{data}
\non_function{}

\title{Survival time of leukemia patients}
\usage{data(whitebloodcell)}
\description{
Survival time of leukemia patients in weeks after diagnosis 
}
\format{
  This data frame contains:
  \describe{
       \item{time:}{survival time of patients after diagnosis in weeks}
       \item{logcount:}{\eqn{\log_10} (initial white blood cell count)}
  }
}
\details{
}
\source{
Cox, D.R. and Snell, E.J. (1981) \emph{Applied Statistics: Principles and Examples},
London: Chapman and Hall, Example U
        }
\references{
Dobson, A. (2002) \emph{ An introduction to Generalized Linear Models},
London: Chapman \& Hall, Exercise: 4.5-4.2
}
\examples{
data(whitebloodcell)
}
\keyword{datasets}
